It should now save images
If the field in the DB is named article_image, it should be saving
images for any new articles that come in.

If it seems to be working, you can clear the articles and posts tables
and run the cron.php file to refill the database if you wanted to do it
for existing entries.

I also stripped their social links and Via links so it'll just be the
content.